requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
Django==1.9.8
django-celery==3.1.17
kombu==3.0.35
mysql-connector-python-rf==2.1.3
pytz==2016.6.1
redis==2.10.3
